+++ src/Home/home.js
@@ -1,5 +1,18 @@
-const HtlmHome = {
-  Home: `<h1>Ola mundo!</h1> `,
-};
+import '../firebase/FireBaseConfig.js';
+import { signinWithGoogle , signinPassword } from '../firebase/authentication.js'
 
+export default() => {
+  const containerHome = document.createElement('div')
 
+  const templateHome = `
+  
+     
+  `; containerHome.innerHTML = template;
+    const textAbout = containerHome.getElementById('text-about');
+    textAbout.addEventListener('click',(e) => {
+        e.preventDefault();
+       const home = document.querySelector('.return-home');
+    })
+
+return container;
+}
