--- src/feed/FeedFunctions.js
@@ -1,1 +1,0 @@
-import '../firebase/FireBaseConfig.js';
